+++ sam/metrics/tests/test_keras_metrics.py
@@ -1,4 +1,5 @@
-from sam.metrics import keras_tilted_loss, keras_rmse, keras_joint_mse_tilted_loss
+from sam.metrics import keras_tilted_loss, keras_rmse, keras_joint_mse_tilted_loss, \
+    keras_joint_mae_tilted_loss
 import numpy as np
 import unittest
 import pytest
@@ -117,6 +118,57 @@ class TestKerasMetrics(unittest.TestCase):
         expected = 0.15
         actual = func([y_true, y_pred])
         self.assertAlmostEqual(expected, actual)
+
+    def test_keras_sum_tilted_loss(self):
+        # We test it with 2 inputs, 1 quantiles, so 4 outputs
+        n_inputs = 2
+        quantiles = [0.1]
+        n_outputs = (len(quantiles) + 1) * n_inputs
+        x = Input(shape=((n_inputs, )))
+        y = Input(shape=((n_outputs, )))
+        func = K.function([x, y], keras_joint_mae_tilted_loss(x, y, quantiles, n_inputs))
+
+        # Just 3 rows
+        y_true = np.array([[1, 2],
+                           [1, 2],
+                           [1, 2]])
+        y_pred = np.array([[0.5, 1.1, 1.1, 2.2],
+                           [0.5, 1.1, 1.1, 2.2],
+                           [0.5, 1.1, 1.1, 2.2]])
+        # These 4 values are `quantile_1_target_1`, `quantile_1_target_2`,
+        # `mean_target_1`, `mean_target_2`
+        # The mean error is 0.1 for the first target, 0.2 for the second target.
+        # So absolute, the mae are 0.1 and 0.2, the sum is 0.3
+        # The mae is 0.5 for the first, 0.9 for the second
+        # But it is quantile, so multiplied by 0.1
+        # So it becomes 0.05 and 0.09. The sum is 0.14
+        # The total sum is 0.3 + 0.5 = 0.44
+        expected = 0.44
+        actual = func([y_true, y_pred])
+
+    def test_keras_sum_tilted_loss_single_input(self):
+        # We test it with 1 inputs, 2 quantiles, so 3 outputs
+        n_inputs = 1
+        quantiles = [0.1, 0.9]
+        n_outputs = (len(quantiles) + 1)
+        x = Input(shape=((n_inputs, )))
+        y = Input(shape=((n_outputs, )))
+        func = K.function([x, y], keras_joint_mae_tilted_loss(x, y, quantiles))
+
+        # Just 3 rows
+        y_true = np.array([[1],
+                           [1],
+                           [1]])
+        y_pred = np.array([[0.5, 0.9, 0.9],
+                           [0.5, 0.9, 0.9],
+                           [0.5, 0.9, 0.9]])
+        # The mean error is 0.1, so absolute also 0.1
+        # The mae is 0.5 for the first quantile, 0.1 for the second
+        # But the first is correct, so gets multiplied by 0.1, so its 0.05
+        # The second is incorrect so gets multiplied by 0.9, so its 0.09
+        # The sum is 0.1 + 0.05 + 0.09 is 0.24
+        expected = 0.24
+        actual = func([y_true, y_pred])
 
 
 if __name__ == '__main__':